+++ src/pocketmine/Player.php
@@ -66,7 +66,6 @@ use pocketmine\event\player\PlayerRespawnEvent;
 use pocketmine\event\player\PlayerToggleFlightEvent;
 use pocketmine\event\player\PlayerToggleSneakEvent;
 use pocketmine\event\player\PlayerToggleSprintEvent;
-use pocketmine\event\player\PlayerConsumeTotemEvent;
 use pocketmine\event\player\PlayerTransferEvent;
 use pocketmine\event\server\DataPacketSendEvent;
 use pocketmine\inventory\CraftingGrid;
@@ -3646,24 +3645,6 @@ class Player extends Human implements CommandSender, ChunkLoader, IPlayer{
 		}
 
 		return false; //never flag players for despawn
-	}
-
-	protected function consumeTotem() : void{
-		$this->server->getPluginManager()->callEvent($ev = new PlayerConsumeTotemEvent($this));
-		if($ev->isCancelled()){
-			$this->kill();
-			return;
-		}
-
-		$this->inventory->setItemInHand(Item::get(Item::AIR));
-
-		$this->broadcastEntityEvent(EntityEventPacket::CONSUME_TOTEM, null, [$this]);
-		$this->level->broadcastLevelEvent($this, LevelEventPacket::EVENT_SOUND_TOTEM);
-
-		$this->removeAllEffects();
-		$this->addEffect(new EffectInstance(Effect::getEffect(Effect::REGENERATION), 800, 1));
-		$this->addEffect(new EffectInstance(Effect::getEffect(Effect::FIRE_RESISTANCE), 800, 1));
-		$this->addEffect(new EffectInstance(Effect::getEffect(Effect::ABSORPTION), 100, 1));
 	}
 
 	protected function respawn() : void{
