+++ src/pocketmine/entity/Entity.php
@@ -429,7 +429,7 @@ abstract class Entity extends Location implements Metadatable, EntityIds{
 	protected $baseOffset = 0.0;
 
 	/** @var float */
-	private $health = 20.0;
+	protected $health = 20.0;
 	private $maxHealth = 20;
 
 	/** @var float */
@@ -930,10 +930,6 @@ abstract class Entity extends Location implements Metadatable, EntityIds{
 	protected function onDeathUpdate(int $tickDiff) : bool{
 		return true;
 	}
-	
-	protected function consumeTotem() : void{
-
-	}
 
 	public function isAlive() : bool{
 		return $this->health > 0;
@@ -958,17 +954,12 @@ abstract class Entity extends Location implements Metadatable, EntityIds{
 
 		if($amount <= 0){
 			if($this->isAlive()){
-				if($this instanceof Player and $this->spawned and $this->hasTotemInHand()){
-					$cause = $this->getLastDamageCause()->getCause();
-					if($cause !== EntityDamageEvent::CAUSE_SUICIDE and $cause !== EntityDamageEvent::CAUSE_VOID){
-					    $this->health = 1;
-						$this->consumeTotem();
-						return;
-					}
+				if($this instanceof Human){
+					$this->playerFinalDamageDestiny();
+				}else{
+					$this->health = 0;
+					$this->kill();
 				}
-
-				$this->health = 0;
-				$this->kill();
 			}
 		}elseif($amount <= $this->getMaxHealth() or $amount < $this->health){
 			$this->health = $amount;
